sdahjk

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -1,7 +1,7 @@
 #include "main.h"
 #include <stdio.h>
 
-/*
+/**
  * puts2: prints every other character of a string
  *
  * @str: the string to compare to output.
@@ -9,12 +9,12 @@
 
 void puts2(char *str)
 {
-	int i = 0;
+	long int i = 0;
 
 	while (str[i] != '\0')
 	{
 		_putchar(str[i]);
-		i+=2;
+		i += 2;
 	}
 
 	_putchar('\n');
